Start the clock on items that are lent rather than given
A thirty-day saddle was arriving with no expiry at all, and the client reads
that as spent: it refuses to use the item and never sends a packet, so the
mount looked broken from the server side with nothing in the log to show for
it. The same was true of every timed item in every shop.

The original stamps an item as it goes into the inventory, not when it is
defined -- PutItem calls SetItemDuration, so the clock starts when the player
receives it. The table says which items run out with a flag of its own at byte
328, confirmed against the whole file: 1450 items carry it and all 1450 carry
a duration, and not one carries a duration without it. Duration is in hours,
and the original adds two on top without saying why.

The encoding is the strange part. There is one two-byte field, Time, at the
end of the item, and TItem.SetExpire writes two different things into it: a
mount counts whole days from a fixed base date, and everything else stores the
unix time shifted right eight bits, three bytes of it, written from one byte
*before* Time. That third byte is the top half of the stack count beside it.
It is not a mistake to tidy up -- the client reads it back the same way -- and
it means a stack of over 255 cannot also expire, which is the original's
problem.

So that an item can be written without the item table in hand everywhere, the
value kept is those three bytes as they sit. A mount's day count is stored
shifted up a byte, leaving a zero where the count's top half is, which changes
nothing because a mount does not stack.

An item bought before this has no expiry and stays unusable; buying it again
gives a good one.

+++ crates/aika-server/src/creation.rs
@@ -691,6 +691,6 @@ fn from_template(item: aika_data::template::Item, container: u8) -> Item {
         durability_min: item.durability_min,
         durability_max: item.durability_max,
         refine: item.refine,
-        expires_at: item.expires_at,
+        expires_at: item.expires_at as u32,
     }
 }
